Change update-to-pinned-version.js to update-dependency.ts
