+++ e2e/scripts/publish-tfjs-ci.sh
@@ -82,4 +82,4 @@ done
 # Update e2e's package.json's all tfjs related packages to locally published
 # version.
 cd e2e
-node ./scripts/update-to-pinned-version.js $RELEASE_VERSION
+yarn update-dependency --version=$RELEASE_VERSION
